fix: allow setting precise resolution for the MJPEG stream (#1009)
* Allow setting precise resolution for the MJPEG stream, by accepting a float instead of an int value for FBMjpegScalingFactor

* Bump version to 9.6.0-repeato.1

* Revert "Bump version to 9.6.0-repeato.1"

This reverts commit e277fb934ad1416b75f82078faec091af3152e62.

+++ WebDriverAgentLib/Utilities/FBConfiguration.h
@@ -136,8 +136,8 @@ extern NSString *const FBSnapshotMaxDepthKey;
  ! Setting this to a value less than 100, especially together with orientation fixing enabled
  ! may lead to WDA process termination because of an excessive CPU usage.
  */
-+ (NSUInteger)mjpegScalingFactor;
-+ (void)setMjpegScalingFactor:(NSUInteger)scalingFactor;
++ (CGFloat)mjpegScalingFactor;
++ (void)setMjpegScalingFactor:(CGFloat)scalingFactor;
 
 /**
  YES if verbose logging is enabled. NO otherwise.
